backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
yfinance
requests
pandas
lxml
python-decouple
curl_cffi
psycopg2-binary
tefas-crawler
